requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
aiosqlite==0.20.0
python-dotenv==1.0.1
google-genai==1.66.0
httpx==0.28.1
apscheduler==3.10.4
python-telegram-bot==21.9
python-multipart==0.0.20
Pillow==11.1.0
